server/requirements.txt
flask
flask-cors
psycopg2-binary
numpy
shapely
pandas
geopandas